데이터 타입 변경 및 오류 수정

+++ server/demo/src/main/java/com/example/demo/Diet/Soup/SoupEntity.java
@@ -21,14 +21,14 @@ public class SoupEntity {
     private String Name;
 
     @Column(nullable = false)
-    private int Calorie;
+    private double Calorie;
 
     @Column(nullable = false)
-    private int Carbohydrate;
+    private double Carbohydrate;
 
     @Column(nullable = false)
-    private int Protein;
+    private double Protein;
 
     @Column(nullable = false)
-    private int Province;
+    private double Province;
 }
